# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Medical Imaging Research Database API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: List patients with pagination and filtering
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: figoStage2018
          in: query
          schema:
            type: string
        - name: pathologyType
          in: query
          schema:
            type: integer
        - name: treatmentOutcome
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Paginated list of patients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientListResponse"
    post:
      operationId: createPatient
      tags: [patients]
      summary: Create a new patient record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePatientBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
  /patients/{id}:
    get:
      operationId: getPatient
      tags: [patients]
      summary: Get patient by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "404":
          description: Not found
    patch:
      operationId: updatePatient
      tags: [patients]
      summary: Update a patient record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePatientBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
    delete:
      operationId: deletePatient
      tags: [patients]
      summary: Delete a patient record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /imaging:
    get:
      operationId: listImagingRecords
      tags: [imaging]
      summary: List imaging records
      parameters:
        - name: patientId
          in: query
          schema:
            type: integer
        - name: modality
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of imaging records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImagingListResponse"
    post:
      operationId: createImagingRecord
      tags: [imaging]
      summary: Create imaging record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateImagingBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImagingRecord"
  /imaging/{id}:
    get:
      operationId: getImagingRecord
      tags: [imaging]
      summary: Get imaging record by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Imaging record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImagingRecord"
    delete:
      operationId: deleteImagingRecord
      tags: [imaging]
      summary: Delete imaging record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /radiomics/features:
    get:
      operationId: listRadiomicsFeatures
      tags: [radiomics]
      summary: List radiomics features for patients
      parameters:
        - name: patientId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Radiomics features list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RadiomicsListResponse"
    post:
      operationId: createRadiomicsFeature
      tags: [radiomics]
      summary: Create radiomics feature record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRadiomicsBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RadiomicsFeature"
  /radiomics/analysis:
    post:
      operationId: runRadiomicsAnalysis
      tags: [radiomics]
      summary: Run radiomics analysis on selected patients
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RadiomicsAnalysisRequest"
      responses:
        "200":
          description: Analysis results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RadiomicsAnalysisResult"
...
  /patients/import:
    post:
      operationId: importPatients
      tags: [patients]
      summary: Bulk import patients from parsed data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                patients:
                  type: array
                  items:
                    $ref: "#/components/schemas/CreatePatientBody"
              required:
                - patients
      responses:
        "201":
          description: Import result
          content:
            application/json:
              schema:
                type: object
                properties:
                  imported:
                    type: integer
                  failed:
                    type: integer
                required:
                  - imported
                  - failed
components:
  schemas:
...
    Patient:
      type: object
      properties:
        id:
          type: integer
        caseNumber:
          type: string
        sequenceNumber:
          type: integer
        patientName:
          type: string
        contactInfo:
          type: ["string", "null"]
        age:
          type: integer
        hypertension:
          type: integer
        diabetes:
          type: integer
        cardiovascular:
          type: integer
        otherComorbidities:
          type: integer
        pregnancyHistory:
          type: ["string", "null"]
        hpvStatus:
          type: ["string", "null"]
        preTreatmentTumorSizeCategory:
          type: ["integer", "null"]
        preTreatmentTumorSize:
          type: ["number", "null"]
        postTreatmentTumorSize:
          type: ["number", "null"]
        tumorSizeChange:
          type: ["number", "null"]
        preSccAg:
          type: ["number", "null"]
        postSccAg:
          type: ["number", "null"]
        preCa125:
          type: ["number", "null"]
        postCa125:
          type: ["number", "null"]
        preCea:
          type: ["number", "null"]
        postCea:
          type: ["number", "null"]
        preCa199:
          type: ["number", "null"]
        postCa199:
          type: ["number", "null"]
        plateletCount:
          type: ["number", "null"]
        lymphocyteCount:
          type: ["number", "null"]
        neutrophilCount:
          type: ["number", "null"]
        monocyteCount:
          type: ["number", "null"]
        serumAlbumin:
          type: ["number", "null"]
        plr:
          type: ["number", "null"]
        lmr:
          type: ["number", "null"]
        pni:
          type: ["number", "null"]
        sii:
          type: ["number", "null"]
        piv:
          type: ["number", "null"]
        preHemoglobin:
          type: ["number", "null"]
        figoStage2009:
          type: ["string", "null"]
        figoStage2018:
          type: ["string", "null"]
        pathologyType:
          type: ["integer", "null"]
        differentiation:
          type: ["string", "null"]
        parametrialInvasion:
          type: ["string", "null"]
        vaginalInvasion:
          type: ["string", "null"]
        pelvicLymphNodeMetastasis:
          type: integer
        inguinalLymphNodeMetastasis:
          type: integer
        commonIliacLymphNodeMetastasis:
          type: integer
        paraAorticLymphNodeMetastasis:
          type: integer
        supraclavicularLymphNodeMetastasis:
          type: integer
        radiationDose:
          type: ["string", "null"]
        brachytherapyDose:
          type: ["string", "null"]
        concurrentChemoCycles:
          type: ["integer", "null"]
        hyperthermia:
          type: integer
        immunotherapy:
          type: integer
        totalTreatmentDays:
          type: ["integer", "null"]
        os:
          type: ["string", "null"]
        pfs:
          type: ["string", "null"]
        treatmentOutcome:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - caseNumber
        - sequenceNumber
        - patientName
        - age
        - hypertension
        - diabetes
        - cardiovascular
        - otherComorbidities
        - pelvicLymphNodeMetastasis
        - inguinalLymphNodeMetastasis
        - commonIliacLymphNodeMetastasis
        - paraAorticLymphNodeMetastasis
        - supraclavicularLymphNodeMetastasis
        - hyperthermia
        - immunotherapy
        - createdAt
    CreatePatientBody:
      type: object
      properties:
        caseNumber:
          type: string
        sequenceNumber:
          type: integer
        patientName:
          type: string
        contactInfo:
          type: ["string", "null"]
        age:
          type: integer
        hypertension:
          type: integer
        diabetes:
          type: integer
        cardiovascular:
          type: integer
        otherComorbidities:
          type: integer
        pregnancyHistory:
          type: ["string", "null"]
        hpvStatus:
          type: ["string", "null"]
        preTreatmentTumorSizeCategory:
          type: ["integer", "null"]
        preTreatmentTumorSize:
          type: ["number", "null"]
        postTreatmentTumorSize:
          type: ["number", "null"]
        tumorSizeChange:
          type: ["number", "null"]
        preSccAg:
          type: ["number", "null"]
        postSccAg:
          type: ["number", "null"]
        preCa125:
          type: ["number", "null"]
        postCa125:
          type: ["number", "null"]
        preCea:
          type: ["number", "null"]
        postCea:
          type: ["number", "null"]
        preCa199:
          type: ["number", "null"]
        postCa199:
          type: ["number", "null"]
        plateletCount:
          type: ["number", "null"]
        lymphocyteCount:
          type: ["number", "null"]
        neutrophilCount:
          type: ["number", "null"]
        monocyteCount:
          type: ["number", "null"]
        serumAlbumin:
          type: ["number", "null"]
        plr:
          type: ["number", "null"]
        lmr:
          type: ["number", "null"]
        pni:
          type: ["number", "null"]
        sii:
          type: ["number", "null"]
        piv:
          type: ["number", "null"]
        preHemoglobin:
          type: ["number", "null"]
        figoStage2009:
          type: ["string", "null"]
        figoStage2018:
          type: ["string", "null"]
        pathologyType:
          type: ["integer", "null"]
        differentiation:
          type: ["string", "null"]
        parametrialInvasion:
          type: ["string", "null"]
        vaginalInvasion:
          type: ["string", "null"]
        pelvicLymphNodeMetastasis:
          type: integer
        inguinalLymphNodeMetastasis:
          type: integer
        commonIliacLymphNodeMetastasis:
          type: integer
        paraAorticLymphNodeMetastasis:
          type: integer
        supraclavicularLymphNodeMetastasis:
          type: integer
        radiationDose:
          type: ["string", "null"]
        brachytherapyDose:
          type: ["string", "null"]
        concurrentChemoCycles:
          type: ["integer", "null"]
        hyperthermia:
          type: integer
        immunotherapy:
          type: integer
        totalTreatmentDays:
          type: ["integer", "null"]
        os:
          type: ["string", "null"]
        pfs:
          type: ["string", "null"]
        treatmentOutcome:
          type: ["integer", "null"]
      required:
        - caseNumber
        - sequenceNumber
        - patientName
        - age
        - hypertension
        - diabetes
        - cardiovascular
        - otherComorbidities
        - pelvicLymphNodeMetastasis
        - inguinalLymphNodeMetastasis
        - commonIliacLymphNodeMetastasis
        - paraAorticLymphNodeMetastasis
        - supraclavicularLymphNodeMetastasis
        - hyperthermia
        - immunotherapy
    UpdatePatientBody:
      type: object
      properties:
        patientName:
          type: string
        contactInfo:
          type: ["string", "null"]
        age:
          type: integer
        hypertension:
          type: integer
        diabetes:
          type: integer
        cardiovascular:
          type: integer
        otherComorbidities:
          type: integer
        pregnancyHistory:
          type: ["string", "null"]
        hpvStatus:
          type: ["string", "null"]
        preTreatmentTumorSizeCategory:
          type: ["integer", "null"]
        preTreatmentTumorSize:
          type: ["number", "null"]
        postTreatmentTumorSize:
          type: ["number", "null"]
        tumorSizeChange:
          type: ["number", "null"]
        preSccAg:
          type: ["number", "null"]
        postSccAg:
          type: ["number", "null"]
        preCa125:
          type: ["number", "null"]
        postCa125:
          type: ["number", "null"]
        preCea:
          type: ["number", "null"]
        postCea:
          type: ["number", "null"]
        preCa199:
          type: ["number", "null"]
        postCa199:
          type: ["number", "null"]
        plateletCount:
          type: ["number", "null"]
        lymphocyteCount:
          type: ["number", "null"]
        neutrophilCount:
          type: ["number", "null"]
        monocyteCount:
          type: ["number", "null"]
        serumAlbumin:
          type: ["number", "null"]
        plr:
          type: ["number", "null"]
        lmr:
          type: ["number", "null"]
        pni:
          type: ["number", "null"]
        sii:
          type: ["number", "null"]
        piv:
          type: ["number", "null"]
        preHemoglobin:
          type: ["number", "null"]
        figoStage2009:
          type: ["string", "null"]
        figoStage2018:
          type: ["string", "null"]
        pathologyType:
          type: ["integer", "null"]
        differentiation:
          type: ["string", "null"]
        parametrialInvasion:
          type: ["string", "null"]
        vaginalInvasion:
          type: ["string", "null"]
        pelvicLymphNodeMetastasis:
          type: integer
        inguinalLymphNodeMetastasis:
          type: integer
        commonIliacLymphNodeMetastasis:
          type: integer
        paraAorticLymphNodeMetastasis:
          type: integer
        supraclavicularLymphNodeMetastasis:
          type: integer
        radiationDose:
          type: ["string", "null"]
        brachytherapyDose:
          type: ["string", "null"]
        concurrentChemoCycles:
          type: ["integer", "null"]
        hyperthermia:
          type: integer
        immunotherapy:
          type: integer
        totalTreatmentDays:
          type: ["integer", "null"]
        os:
          type: ["string", "null"]
        pfs:
          type: ["string", "null"]
        treatmentOutcome:
          type: ["integer", "null"]
    PatientListResponse:
      type: object
      properties:
        patients:
          type: array
          items:
            $ref: "#/components/schemas/Patient"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - patients
        - total
        - page
        - totalPages
...
    ImagingRecord:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patientName:
          type: string
        modality:
          type: string
        bodyPart:
          type: string
        studyDate:
          type: string
          format: date-time
        description:
          type: ["string", "null"]
        findings:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - patientId
        - patientName
        - modality
        - bodyPart
        - studyDate
        - createdAt
    CreateImagingBody:
      type: object
      properties:
        patientId:
          type: integer
        modality:
          type: string
        bodyPart:
          type: string
        studyDate:
          type: string
          format: date-time
        description:
          type: ["string", "null"]
        findings:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
      required:
        - patientId
        - modality
        - bodyPart
        - studyDate
    ImagingListResponse:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: "#/components/schemas/ImagingRecord"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - records
        - total
        - page
        - totalPages
...
    RadiomicsListResponse:
      type: object
      properties:
        features:
          type: array
          items:
            $ref: "#/components/schemas/RadiomicsFeature"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - features
        - total
        - page
        - totalPages
...